moved generate_service_extras
